Commiting to merge with feature/xml-menu

+++ Components/PhysicsComponent.hpp
@@ -4,25 +4,24 @@
 #include "Component.hpp"
 
 class PhysicsComponent : public Component {
-public:
-    string wow = "huseyin";
-
 private:
     const PhysicsAPI *enginePhysicsAPI;
-    unsigned int bodyId;
+    unsigned int bodyId = 0;
 
 public:
-    explicit PhysicsComponent() : Component(), bodyId(0) {
-
-    }
+    explicit PhysicsComponent(EntityId id);
 
     PhysicsComponent(EntityId id, BodyType bodyType, Vector2 position, Vector2 size);
 
     PhysicsComponent(EntityId id, BodyType bodyType, Vector2 position, float radius);
 
-    unique_ptr<Component> Clone(EntityId entityId) override;
+    [[nodiscard]] Component *Clone(EntityId entityId) const override;
 
-    ~PhysicsComponent(){
+    ~PhysicsComponent() override {
+        // BodyId 0 is used for physicsComponent that are not instantiated within the b2World.
+        if(bodyId != 0)
+            return;
+
         enginePhysicsAPI->destroyBody(bodyId);
     }
 
@@ -30,24 +29,24 @@ public:
         return enginePhysicsAPI->getRPosition(bodyId);
     }
 
-    inline void getVelocity(Vector2& velocity) {
+    inline void getVelocity(Vector2 &velocity) {
         enginePhysicsAPI->GetVelocity(velocity, bodyId);
     }
 
-    inline void setVelocity(const Vector2& velocity) {
+    inline void setVelocity(const Vector2 &velocity) {
         enginePhysicsAPI->setLinearVelocity(bodyId, velocity);
     }
 
-    inline void setFixedRotation(bool value)
-    {
+    inline void setFixedRotation(bool value) {
         enginePhysicsAPI->setFixedRotation(bodyId, value);
     }
+
+    [[nodiscard]] std::string name() const override;
 
     void update() override;
 
 private:
-    inline BodyId initializeBoxBody(BodyType bodyType, Vector2 position, Vector2 size)
-    {
+    inline BodyId initializeBoxBody(BodyType bodyType, Vector2 position, Vector2 size) {
         Box2DBoxData box2DBoxData;
         box2DBoxData.bodyType = bodyType;
         box2DBoxData.position = position;
@@ -55,8 +54,7 @@ private:
         return enginePhysicsAPI->createStaticBody(box2DBoxData);
     }
 
-    inline BodyId initializeCircleBody(BodyType bodyType, Vector2 position, float radius)
-    {
+    inline BodyId initializeCircleBody(BodyType bodyType, Vector2 position, float radius) {
         Box2DCircleData box2DBoxData;
         box2DBoxData.bodyType = bodyType;
         box2DBoxData.position = position;
